Fixed most warnings when compiling with Visual Studio 2019 all the way up to warning level 4 and with conformance mode on (progress on #24).
There are many warnings that I didn't bother with fixing, because they're either just suggested changes that we don't need to bother with, such as using unscoped enums as opposed to enum class (which can't be fully fixed anyways, because those are used in the C library headers, such as SDL.h), or false positives due to the compiler not being able to figure out there's nothing wrong. There are also some warnings that can't be easily fixed, due to the offending code being very C-style, and Visual Studio not properly analyzing said code.

More testing with other compilers/platforms needs to be done to be sure no more warnings worth fixing remain.

+++ src/SGUIL/SGUIL.cc
@@ -328,7 +328,7 @@ void Gui_GenerateTextPositionalValuesPartial(std::string& text, unsigned int pos
 
             if(i != 0 || text[i] == '\n')
             {
-                currentY += fmt->lineSpacing * fmt->sizeMult * (float)font.charH;
+                currentY += static_cast<int>(fmt->lineSpacing * fmt->sizeMult * (float)font.charH);
             }
 
             unsigned int width = (unsigned int)((float)(destBox.w) / (float)(font.charW) / fmt->sizeMult);
@@ -358,7 +358,7 @@ void Gui_GenerateTextPositionalValuesPartial(std::string& text, unsigned int pos
                 case enumAlignment::center:
                     if(lines[linefeeds].size() - lastWrapLinePos < width)
                     {
-                        currentX = destBox.x + (fmt->sizeMult * (float)font.charW / 2.0) * (width - (lines[linefeeds].size() - lastWrapLinePos));
+                        currentX = static_cast<int>(destBox.x + (fmt->sizeMult * (float)font.charW / 2.0f) * (width - (lines[linefeeds].size() - lastWrapLinePos)));
                     } else
                     {
                         currentX = destBox.x;
@@ -382,7 +382,7 @@ void Gui_GenerateTextPositionalValuesPartial(std::string& text, unsigned int pos
 
         values.push_back( {currentX, currentY} );
 
-        currentX += fmt->sizeMult * (float)font.charW;
+        currentX += static_cast<int>(fmt->sizeMult * (float)font.charW);
     }
 }
 
@@ -454,8 +454,8 @@ void Gui_DrawTextPartial_PV(std::string text, unsigned int pos, unsigned int len
         }
 
         if(fmt->shadow) {
-            dest.x -= 2.0 * fmt->sizeMult;
-            dest.y += 2.0 * fmt->sizeMult;
+            dest.x -= static_cast<int>(2.0f * fmt->sizeMult);
+            dest.y += static_cast<int>(2.0f * fmt->sizeMult);
 
             SDL_SetTextureAlphaMod(font.sheet, (Uint8)((float)rgba_A(fmt->rgba) / 3.5));
             if(font.outlineSheet)
@@ -467,8 +467,8 @@ void Gui_DrawTextPartial_PV(std::string text, unsigned int pos, unsigned int len
                 SDL_RenderCopy(Gui_SDL_Renderer, font.outlineSheet, &src, &dest);
             }
 
-            dest.x += 2.0 * fmt->sizeMult;
-            dest.y -= 2.0 * fmt->sizeMult;
+            dest.x += static_cast<int>(2.0f * fmt->sizeMult);
+            dest.y -= static_cast<int>(2.0f * fmt->sizeMult);
 
             SDL_SetTextureAlphaMod(font.sheet, rgba_A(fmt->rgba));
             if(font.outlineSheet)
